+++ src/main/java/hotel/domain/repository/GuestRepository.java
@@ -1,6 +1,6 @@
 package hotel.domain.repository;
 
-import hotel.data.entity.Guest;
+import hotel.data.entity.guest.Guest;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -8,7 +8,7 @@ import java.util.Optional;
 import java.util.UUID;
 
 @Repository
-public interface GuestRepository extends JpaRepository<Guest, Integer> {
+public interface GuestRepository extends JpaRepository<Guest, Long> {
 
     Optional<Guest> findByRg(String rg);
     Optional<Guest> findByRgOrDocument(String rg, String document);
